SBERT title embeddings
